add heading in accordion theme
