+++ src/js/components/AccordionPanel/doc.js
@@ -16,6 +16,11 @@ export function doc(Panel) {
 }
 
 export const themeDoc = {
+  'accordion.heading.level': {
+    description: 'The heading level used for the accordion.',
+    type: 'number',
+    defaultValue: '4',
+  },
   'accordion.icons.collapse': {
     description: 'The icon to use when the panel is expanded.',
     type: 'React.Element',
