+++ src/js/themes/base.js
@@ -260,14 +260,15 @@ export const generate = (baseSpacing = 24, scale = 6) => {
       },
     },
     accordion: {
+      border: {
+        side: 'bottom',
+        color: 'border',
+      },
+      heading: { level: '4' }, // level ranges from 1-6
       icons: {
         collapse: FormUp,
         expand: FormDown,
         // color: { dark: undefined, light: undefined },
-      },
-      border: {
-        side: 'bottom',
-        color: 'border',
       },
     },
     anchor: {
